data_ingestion/requirements.txt
telethon==1.32.0
sqlalchemy==2.0.23
beautifulsoup4==4.12.2
requests==2.28.2
openai==1.12.0
apscheduler==3.10.4
python-dotenv==1.0.0
click==8.1.7
selenium==4.15.0
webdriver-manager==4.0.1
